Move withParams to AbstractUrl
